Delete associated records when deleting user

+++ server/routes/register.js
@@ -3,6 +3,7 @@ const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const User = require('../models/user');
 const Token = require('../models/token');
+const URL = require('../models/url');
 
 router.post('/', async (req, res, next) => {
     try {
@@ -32,13 +33,33 @@ router.delete('/', async (req, res, next) => {
 
     try {
         const { userID } = jwt.verify(cookie, process.env.JWT_REFRESH_SECRET);
-        const user = await User.findByIdAndRemove(userID);
+
+        if (!req.body.username || !req.body.password) {
+            res.status(400);
+            throw new Error('Missing username or password');
+        }
+
+        const user = await User.findOne({ username: req.body.username });
+
+        if (user === null) {
+            res.status(400);
+            throw new Error('Invalid username or password');
+        }
+
+        if (!(await bcrypt.compare(req.body.password, user.password))) {
+            res.status(400);
+            throw new Error('Invalid username or password');
+        }
+
+        await user.deleteOne();
         const token = await Token.findByIdAndRemove(userID);
 
         if (user === null && token === null) {
             res.status(400);
             throw new Error("User doesn't exist");
         }
+
+        await URL.deleteMany({ userID: user._id });
 
         res.clearCookie('refreshToken');
         res.json({ message: `User ${user.username} deleted` });
